+++ src/module/data/pseudo-documents/pseudo-document.mjs
@@ -294,7 +294,7 @@ export default class PseudoDocument extends foundry.abstract.DataModel {
    * @param {object} operation                              The context of the operation.
    * @param {foundry.abstract.DataModel} operation.parent   The parent of this document.
    * @param {boolean} [operation.renderSheet]               Render the sheet of the created pseudo-document?
-   * @returns {Promise<foundry.abstract.Document>}          A promise that resolves to the updated document.
+   * @returns {Promise<PseudoDocument>} A promise that resolves to the created PseudoDocument.
    */
   static async create(data = {}, { parent, renderSheet = true, ...operation } = {}) {
     if (!parent) {
@@ -312,8 +312,9 @@ export default class PseudoDocument extends foundry.abstract.DataModel {
     const update = { [`${fieldPath}.${id}`]: { ...data, _id: id } };
     this._configureUpdates("create", parent, update, operation);
     await parent.update(update, operation);
-    if (renderSheet) parent.getEmbeddedDocument(this.metadata.documentName, id).sheet?.render({ force: true });
-    return parent;
+    const pseudo = parent.getEmbeddedDocument(this.metadata.documentName, id);
+    if (renderSheet) pseudo.sheet?.render({ force: true });
+    return pseudo;
   }
 
   /* -------------------------------------------------- */
@@ -323,7 +324,7 @@ export default class PseudoDocument extends foundry.abstract.DataModel {
    * @param {object} [data]                                 The data used for the creation.
    * @param {object} operation                              The context of the operation.
    * @param {foundry.abstract.Document} operation.parent    The parent of this document.
-   * @returns {Promise<foundry.abstract.Document|null>}     A promise that resolves to the updated document.
+   * @returns {Promise<PseudoDocument|null>}     A promise that resolves to the updated document.
    */
   static async createDialog(data = {}, { parent, ...operation } = {}) {
     // If there's demand or need we can make the template & context more dynamic